Split ops into ops.op and ops.builtin
