+++ docs/source/ops.rst
@@ -1,6 +1,13 @@
 Operations
 ----------
-.. automodule:: funsor.ops
+
+.. automodule:: funsor.ops.op
+    :members:
+    :undoc-members:
+    :show-inheritance:
+    :member-order: bysource
+
+.. automodule:: funsor.ops.builtin
     :members:
     :undoc-members:
     :show-inheritance:
