+++ funsor/ops/__init__.py
@@ -1,0 +1,3 @@
+from .array import *
+from .builtin import *
+from .op import *
