+++ setup.cfg
@@ -5,6 +5,7 @@ ignore = F811,E121,E123,E126,E226,E24,E704,W503,W504
 per-file-ignores =
     test/examples/test_bart.py:E128
     test/examples/test_sensor_fusion.py:E128
+    funsor/ops/__init__.py:F401,F403
 
 [isort]
 line_length = 120
